+++ rviz_default_plugins/src/rviz_default_plugins/displays/tf/tf_display.hpp
@@ -27,8 +27,8 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef RVIZ_DEFAULT_PLUGINS__DISPLAYS__TF_DISPLAY__TF_DISPLAY_HPP_
-#define RVIZ_DEFAULT_PLUGINS__DISPLAYS__TF_DISPLAY__TF_DISPLAY_HPP_
+#ifndef RVIZ_DEFAULT_PLUGINS__DISPLAYS__TF__TF_DISPLAY_HPP_
+#define RVIZ_DEFAULT_PLUGINS__DISPLAYS__TF__TF_DISPLAY_HPP_
 
 #include <map>
 #include <memory>
@@ -161,4 +161,4 @@ private:
 }  // namespace displays
 }  // namespace rviz_default_plugins
 
-#endif  // RVIZ_DEFAULT_PLUGINS__DISPLAYS__TF_DISPLAY__TF_DISPLAY_HPP_
+#endif  // RVIZ_DEFAULT_PLUGINS__DISPLAYS__TF__TF_DISPLAY_HPP_
